third_party/WebKit/LayoutTests/fast/dom/StyleSheet/empty-shadow-style.html

<!doctype html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<div id='sandbox'></div>
<pre id='console'></pre>
<script>
var sandbox = document.getElementById('sandbox');
var host = document.createElement('div');
var shadowRoot = host.createShadowRoot();
shadowRoot.innerHTML = '<style></style><div>Shadow</div>';
shadowRoot.appendChild(document.createElement('style'));
sandbox.appendChild(host);

shouldBe("shadowRoot.styleSheets.length", "2");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
